requirements.txt
boto3>=1.34.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-monitor>=6.0.2
azure-mgmt-containerservice>=28.0.0
azure-mgmt-web>=7.0.0
azure-mgmt-cosmosdb>=9.0.0
azure-mgmt-sql>=3.0.0
fastapi
uvicorn
python-dotenv
openai>=1.0.0
pydantic
sqlalchemy
python-multipart
streamlit>=1.32.0
streamlit-option-menu>=0.3.12
plotly>=5.20.0
